[TASK] Update phpunit to versions without prophecy requirement
phpspec/prophecy was a hard dependency for phpunit, which has
changed in the latest versions of 8.5 and 9.5. phpspec/prophecy
needs to ignore php for platform requirement checks with PHP8.2

This change updates the dev dependency for phpunit to the latest
versions which do not include the hard requirement anymore. This
allows installation of phpunit with PHP8.2.

Used command(s):

```shell
composer require --dev --no-update \
  "phpunit/phpunit":"^7.5.20 || ^8.5.30 || 9.5.25"
```